+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Version 1.1.0
 * added extra reservoir
+* added extra hex
+* added extra arcanist exploit
+* added extra arcana
+
+* added extra rogue talent
+* added extra slayer talent
 
 ## Version 1.0.9
 * added focused weapon advanced weapon training