+++ CHANGELOG.md
@@ -4,10 +4,10 @@
 * added extra hex
 * added extra arcanist exploit
 * added extra arcana
-
 * added extra rogue talent
 * added extra slayer talent
 * added extra revelation
+* added extra discovery
 
 ## Version 1.0.9
 * added focused weapon advanced weapon training